chore(audio): split demo
