+++ src/packages/__VUE/audio/demo/basic.vue
@@ -1,0 +1,6 @@
+<template>
+  <nut-audio :url="url" loop type="icon"></nut-audio>
+</template>
+<script setup lang="ts">
+const url = '//storage.360buyimg.com/jdcdkh/SMB/VCG231024564.wav';
+</script>
