+++ src/packages/__VUE/audio/doc.md
@@ -16,129 +16,19 @@ app.use(Audio);
 
 ### 基础用法
 
-:::demo
-
-```vue
-<template>
-  <nut-audio url="//storage.360buyimg.com/jdcdkh/SMB/VCG231024564.wav" :loop="true" type="icon"></nut-audio>
-</template>
-```
-
-:::
+> demo: audio basic
 
 ### 语音播放
 
-:::demo
-
-```vue
-<template>
-  <nut-audio url="//storage.360buyimg.com/jdcdkh/SMB/VCG231024564.wav" :loop="false" type="none" ref="audioDemo">
-    <div class="nut-voice">
-      <div><Voice></Voice></div>
-      <div>{{ duration }}"</div>
-    </div>
-  </nut-audio>
-</template>
-<script setup>
-import { ref, onMounted } from 'vue';
-import { Voice } from '@nutui/icons-vue';
-const audioDemo = ref(null);
-const duration = ref(0);
-onMounted(() => {
-  setTimeout(() => {
-    duration.value = audioDemo.value.second.toFixed();
-  }, 500);
-});
-</script>
-
-<style>
-.nut-voice {
-  display: flex;
-  justify-content: space-between;
-  width: 100px;
-  height: 20px;
-  padding: 8px;
-  border: 1px solid rgba(0, 0, 0, 0.6);
-  border-radius: 18px;
-}
-</style>
-```
-
-:::
+> demo: audio play
 
 ### 进度条展示
 
-:::demo
-
-```vue
-<template>
-  <nut-audio url="//storage.360buyimg.com/jdcdkh/SMB/VCG231024564.wav" :loop="true" type="progress">
-    <div class="nut-audio-operate-group">
-      <nut-audio-operate type="back"></nut-audio-operate>
-      <nut-audio-operate type="play"></nut-audio-operate>
-      <nut-audio-operate type="forward"></nut-audio-operate>
-      <nut-audio-operate type="mute"></nut-audio-operate>
-    </div>
-  </nut-audio>
-</template>
-```
-
-:::
+> demo: audio progress
 
 ### 自定义操作按钮
 
-:::demo
-
-```vue
-<template>
-  <nut-audio
-    url="//storage.360buyimg.com/jdcdkh/SMB/VCG231024564.wav"
-    :loop="false"
-    type="progress"
-    @forward="forward"
-    @fastBack="fastBack"
-    @play="changeStatus"
-    @ended="ended"
-    @changeProgress="changeProgress"
-  >
-    <div class="nut-audio-operate-group">
-      <nut-audio-operate type="back">
-        <PlayDoubleBack width="35px" height="35px"></PlayDoubleBack>
-      </nut-audio-operate>
-      <nut-audio-operate type="play">
-        <PlayStart v-if="!playing" width="35px" height="35px"></PlayStart>
-        <PlayStop v-else width="35px" height="35px"></PlayStop>
-      </nut-audio-operate>
-      <nut-audio-operate type="forward">
-        <PlayDoubleForward width="35px" height="35px"></PlayDoubleForward>
-      </nut-audio-operate>
-    </div>
-  </nut-audio>
-</template>
-<script setup>
-import { ref } from 'vue';
-import { PlayDoubleBack, PlayDoubleForward, PlayStart, PlayStop } from '@nutui/icons-vue';
-const playing = ref(false);
-const fastBack = () => {
-  console.log('倒退');
-};
-const forward = (progress) => {
-  console.log('快进', '当前时间' + progress);
-};
-const changeStatus = (status) => {
-  console.log('当前播放状态', status);
-  playing.value = status;
-};
-const ended = () => {
-  console.log('播放结束');
-};
-const changeProgress = (val) => {
-  console.log('改变进度条', val);
-};
-</script>
-```
-
-:::
+> demo: audio control
 
 ## API
